refactor(pools): reorganize the src folder structure
