+++ src/base-classes/base-provider.ts
@@ -1,5 +1,5 @@
-import { ERC20CollateralPool } from './erc20-collateral-pool'
-import { Pools as PoolsClass } from './pools'
+import { ERC20CollateralPool } from '../pools/erc20-collateral-pool'
+import { Pools as PoolsClass } from '../pools/pools'
 
 export abstract class BaseProvider<T extends PoolsClass | ERC20CollateralPool> {
   readonly contract: T
