+++ test/integration/self-provider-erc20collateralpool.ts
@@ -1,9 +1,9 @@
-import { Erc20 } from '../../src/erc20'
-import { ERC20CollateralPool } from '../../src/erc20-collateral-pool'
-import { ecpErrorMessage } from '../../src/error-messages'
-import { SelfProvider } from '../../src/self-provider'
+import { ecpErrorMessage } from '../../src/errors/error-messages'
+import { ERC20CollateralPool } from '../../src/pools/erc20-collateral-pool'
+import { SelfProvider } from '../../src/pools/self-provider'
 import { Lend, Pool } from '../../src/types/erc20-collateral-token'
-import { sleep } from '../../src/util'
+import { Erc20 } from '../../src/utilities/erc20'
+import { sleep } from '../../src/utilities/util'
 import {
   ADMIN_TESTING_PRIVATE_KEY,
   COLLATERAL_TOKEN,
